feat(freeGame): Handle announcement + refacto
